feat: shared JSON-RPC transport
